+++ CHANGES.md
@@ -5,8 +5,9 @@ Changes in 3.7-beta
 -------------------
 - RTL layout issues #214 in rtl.css
 - INSERT INTO case to Replace empty strings ('') with NULL values in database.inc.php
-- Add DISPLAY_NAME to CONFIG table in rosariosis.sql & Update.fnc.php
+- #218 Add DISPLAY_NAME to CONFIG table in rosariosis.sql & Update.fnc.php
 - Fix #221 Delete Addresses & Contacts info does not delete student in Student.php & Address.inc.php
+- #218 Add getDisplayNameSQL() & getDisplayName() functions in GetStuList.fnc.php
 
 Changes in 3.6.1
 ----------------